requirements.txt
flask
sqlalchemy
requests
python-dotenv
openai
pandas
ebooklib
jupyter